Consolidate shared code between menu item types
